Updating to version 196 - Fixing world hopping interface numbers - adding better dialog status

+++ src/com/p3achb0t/detours/Detours.java
@@ -15,7 +15,7 @@ public class Detours {
                                 String menuTarget,
                                 int mouseX,
                                 int mouseY,
-                                int dummy,
+                                byte dummy,
                                 InstanceManager scriptManager){
 
         ActionScript script = scriptManager.getActionScript();
